+++ app/src/main/java/com/wyb/wyb_android/ui/calendar/CalendarViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import com.wyb.wyb_android.data.model.Discomfort
-import java.text.SimpleDateFormat
+import com.wyb.wyb_android.util.Utils
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.util.*
@@ -45,7 +45,7 @@ class CalendarViewModel : ViewModel() {
 
     private fun formatDates() {
         for (dateString in eventDates) {
-            val localDate = convertIsoStringToLocalDate(dateString)
+            val localDate = Utils.convertIsoStringToLocalDate(dateString)
             if (localDate != null) {
                 localDates.add(arrayListOf(localDate))
             }
@@ -133,19 +133,6 @@ class CalendarViewModel : ViewModel() {
         }
     }
 
-    private fun convertIsoStringToLocalDate(isoDate: String): LocalDate? {
-        // TODO: 서버 연동 후 timeZone Korea 로 변경
-        val sdf = SimpleDateFormat(ISO_DATE_FORMAT, Locale.getDefault())
-        val calendar = Calendar.getInstance()
-        calendar.time = sdf.parse(isoDate) ?: return null
-        calendar.timeZone = TimeZone.getTimeZone("Etc/UTC")
-        return LocalDate.of(
-            calendar.get(Calendar.YEAR),
-            calendar.get(Calendar.MONTH) + 1,
-            calendar.get(Calendar.DAY_OF_MONTH)
-        )
-    }
-
     private fun convertCalendarDayToLocalDate(calDay: CalendarDay): LocalDate? {
         return LocalDate.of(calDay.year, calDay.month, calDay.day)
     }
@@ -215,9 +202,5 @@ class CalendarViewModel : ViewModel() {
                 }
             }
         }
-    }
-
-    companion object {
-        private const val ISO_DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss.SSSXXX"
     }
 }
